+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Add Docker
 + Get public IP
++ Promisify getZoneId
 
 ## 0.1.0
 + First version - forked